IOS-3449: Extract InfinityFrame helper into separate file
